CLN: fix review
